Update trace and span id conversion from Long to String.
When sending trace and span ids over the wire Zipkin hex encodes
the id's in case http is used. There was an earlier fix for Brave
which made the hex encoding compatible with Zipkin / Finagle but this
fix was not complete. Negative span and trace id's were not compatible.
This commit fixes this.

This is important in case you want to mix brave services with finagle
zipkin services.