Flattens Impl classes with AutoValue and Removes commons-lang3
In many cases, immutable class trees end in an only child leave,
separating an interface from its only implementation. In other cases,
an Impl suffix was present on types that are unambiguous.

This change flattens some of this, employing google AutoValue. The
result is a net reduction in complexity and lines of code.

This will break some folks who depend directly on Impl classes or their
constructors. However, 3.x release is already breaking and accomodating
these changes is largely mechanical.

Finally, this removes commons-lang3 using AutoValue and some inlining.