Fix and re-enable integration tests
It looks like the root cause for the integration test failures has been found.
Requesting our SpanProcessingThread to stop closed the thrift client. This
thrift client was in some cases still in use to send spans. As a result an exception
was thrown which resulted in resending the last batch of spans with a new client.
This means we had 'at least once' delivery and not 'exactly once' as the test expected.

The client is closed now in ZipkinSpanCollector, who also creates the client, not in SpanProcessingThread.

There is also some clean-up of the tests and each test has its own test class now.