Adds Span.timestamp, duration and address annotations
This updates Brave to the latest thrift specification, which includes
explicit settings for Span timestamp and duration, as well as binary
annotations that indicate the actual IP of a client or server.

This also updates to lowercase convention for Span.name and
Endpoint.serviceName, also noted in recent thrift documentation.