Adds http.url and http.status_code to http request adapters
Problem:
When hitting http://example.com/a/b/c with a browser,
the annotations http.uri and http.responsecode are never recorded.
This means we will not be able to figure out where on the
web site a trace was triggered.

Solution:
Add annotations to server request/response adapters

Zipkin uses the TraceKeys-class to ensure consistent keys. This commit
updates Brave to use the same TraceKeys-class to ensure future compatibility.