Adds internal CurrentSpan type and deprecates SpanAndEndpoint (#318)
SpanAndEndpoint was only exposed for a historical method factory on
AnnotationSubmitter. In fact, `Brave.annotationSubmitter()` is
functionally equivalent to `Brave.serverTracer()`. This change
deprecates both the redundant `Brave.annotationSubmitter()` method
and also the `SpanAndEndpoint` type.

In its place are the following:

* new `CurrentSpan` type, retrofitted to the thread binders. This is
  not exposed publicly, as there's no reason to at this point.
* eager resolution of the localEndpoint (which is currently passed
  into the thread state classes). In almost all cases, this will have
  no effect since constructors eagerly resolve this anyway.

These changes make it more clear what's going on, particularly around
logic that decides whether to create a new trace or not. It also makes
replacing logic simpler, which is needed for Brave 4.