Backfills tests for adding remote endpoint (#394)
Before, we were inconsistent on whether or not to do an nslookup when
getting the remote endpoint address. This ensures we never do. It also
enforces a policy that we don't add an address tag with no IP and no
service name. Most importantly, this backfills tests for all related
logic.