Adds HttpSampler for parameterized sampling of http requests (#408)
The prior sampling policy was to use the default (trace ID) sampler for
server and client requests.  For example, if there's a incoming request
that has no trace IDs in its headers, the sampler indicated by
`Tracing.Builder.sampler` decides whether or not to start a new trace.
Once a trace is in progress, it is used for any outgoing http client
requests.

On the other hand, you may have http client requests that didn't originate
from a server. For example, you may be bootstrapping your application,
and that makes an http call to a system service. The default policy will
start a trace for any http call, even ones that didn't come from a server
request.

This change makes it possible to make sampling decisions based on the
http request, in a way that applies to all http instrumentation uniformly.

Here's an example which only starts traces for requests
originating from the  `/api` endpoint. It chooses 100% of such requests.

```java
httpTracing = httpTracing.toBuilder()
    .serverSampler(new HttpSampler() {
       @Override public <Req> Boolean trySample(HttpAdapter<Req, ?> adapter, Req request) {
         return adapter.path(request).startsWith("/api");
       }
     })
    .clientSampler(HttpSampler.NEVER_SAMPLE)
    .build();
```