Replaces Tagger with SpanCustomizer for user callbacks (#417)
* Replaces Tagger with SpanCustomizer for user callbacks

Before, we introduced `brave.Tagger` to limit the scope when adding tags
to spans. This widens the scope a bit to handle all routine changes made
by users to in-flight spans:

* setting the span name
* adding tags
* adding annotations

This type is called `SpanCustomizer`, and only exposes the above rather
than lifecycle hooks such as `start()`. This has the following benefits:

* Users do not need to learn the entire span type, or be tempted to finish one!
* Users can prevent accidental huge tags by by decorating the `SpanCustomizer`

Thanks to @devinsba for the idea. Now's the best time to change this as
we haven't released any of the affected code formally.

* changed back to interface