Tests and documents SparkJava as compatible with servlet deployment
There are a number of examples on the web where SparkJava runs in a war
file or in a Spring Boot application. This documents to use the normal
`TracingFilter` in this case.

Note: sometime in the future, we may add a span customizing feature
which could publish SparkJava's route information to the servlet layer.