Renames "b3SingleFormat" to "writeB3SingleFormat" and simplifies b3 impl
This renames the setting "b3SingleFormat" to "writeB3SingleFormat" as
it only applies to outbound propagation. It also simplifies the "b3"
implementation by not checking for two keys (the old impl was also case
sensitive by default). This allows us to special case when only "b3" is
used. Finally, this backfills some JMS related tests and benchmarks.