Ensures we don't apply assembly scope on subscription callbacks
Before, we were inconsistent about whether we apply the assembly context
to subscription callbacks. My current understanding is that this is not
a good idea, as subscribing is different than value publication or
consumption. Accordingly, this makes everything consistent in so far as
calls to `subscribe` or `onSubscribe` are left alone, meaning the trace
context calling `subscribe` will not be affected in any way.

Note: there is some nuance which is that when a source is backed by a
`Callable`, callbacks such as `onNext` as invoked directly from thei
subscription thread. We still apply the assembly trace context to these
callbacks for consistency.