Adds HttpRequestSampler and backports HttpSampler to use it (#994)
We recently introduced the `HttpClientRequest` and `HttpServerRequest`
types. These are used during propagation extraction, for example as a
parameter to `Extractor.extract(request)`. This was to facilitate
secondary sampling, which occurs just prior to the primary sampling as
performed by `HttpSampler`.

One feedback was that it is intuitive to use types like
`HttpRuleSampler` uniformly across primary and secondary sampling.

This change introduces `HttpRequestSampler` which aligns types so that
rules can be defined with similar code regardless of primary or
secondary decision. This also takes care to not break api. In order to
reduce complexity around starting a span with old or new types, this
pushes some logic to adapters. Rationale is adapters are straightforward
to test and easier to reason with than other logic.

See https://github.com/openzipkin-contrib/zipkin-secondary-sampling