Introduces `Response.error()` and uses it in HTTP abstraction (#1085)
This introduces `Response.error()`, present when there was an exception
invoking a client request, or processing a server request. This also
ensures that `FinishedSpanHandler` can read it independently of the
"error" tag.

This implied retro-fitting this in the HTTP abstraction and backfilling
tests that tell the difference between HTTP status and a parsed
exception. This identified a gap in Servlet libraries which formerly
swallowed application level exceptions. To bridge that gap, we now
look at the request attribute "error" when ending a server span.

Co-authored-by: Anuraag Agrawal <anuraaga@gmail.com>