Makes StrictScopeDecorator stricter without renaming it (#1127)
This fixes some fuzz in tests, but notably makes `StrictScopeDecorator`
`Closeable`. In doing so, this will expose any instrumentation that
leaked scopes as opposed to before which could only show if something
closed on the wrong thread.

This type was only ever documented for development, so there should be
no harm in updating this in a way that could OOM in extreme leak
circumstances (due to hard references).