Uses SpanCustomizer as a base for all running spans (#1134)
This simplifies concepts and code by retrofitting `ScopedSpan` as a
`SpanCustomizer`. In doing so, it makes hierarchies easier as everything
that can affect a current span can use a `SpanCustomizer` to do so.

This implicitly allows renaming of a `ScopedSpan` which can be done
anyway via `Tracer.currentSpan()`

Thanks to @anuraaga for the idea.