Adds BaggagePropagation.allKeyNames and deprecates non-string propagation (#1171)
In 5.11, we partially replaced functionality of ExtraFieldPropagation.extraKeys().
This completes the work after I noticed it was needed for several use cases, not
just OpenTracing. `BaggagePropagation.allKeyNames()` is also more robust when
propagation is wrapped by proxy or otherwise.

The Propagation `<K>` parameter allowed integration with requests that had
headers, but didn't use String types. This was over-generalized, as in practice
only gRPC instrumentation ever used this key type (`Metadata.Key`).

Removing this `<K>` parameter dramatically simplifies the model as it removes
the need to explain the key factory and the edge case it supported, which can
be accomplished differently.

In Brave 6, we will need to keep the generic parameter `Propagation<String>`
(to avoid API breaks). However, the internals will be far less complex once we
can remove KeyFactory etc that are no longer needed.