Consistently only clear trace ID headers in messaging (#1229)
Before, we cleared only trace ID headers in messaging, but cleared all
headers in Kafka and Rabbit. The problem in the latter was that it would
clear baggage like "user_id" and "country_code". These don't interfere
with the trace tree and worse, can prevent users from seeing correlation
IDs in third party code that doesn't use Brave.

The fix here is to clarify code and docs, moving implementation of Kafka
and RabbitMQ to match JMS.